requirements.txt
torch
torchvision
torchaudio
transformers
peft
datasets
accelerate
fastapi
uvicorn
pillow
scikit-learn
evaluate
python-dotenv
